requirements.txt
playwright
python-dotenv
pandas
httpx
fastapi
uvicorn

